bar1 a2a round loop: a finished block keeps its offset inside the tensor; LSE merge names the bf16 wire
fn7j (19.09.): the uneven LSE-merge all_to_all (blocks 15/5/4 heads x 8192
tokens, fp32) died on every rank in the first prefill chunk with
'barlink-bar1 a2a: send block 2 (102707200+0) lies past the end of the
input tensor (100663296 bytes)'. The multi-round loop in
barlink_all_to_all_single advanced every block's offset by k*slot, also
for a block that had already finished (length 0), and the extension's
bounds check (off + len <= in_bytes) refuses that. Offset now
base + min(k*slot, length). First real multi-round uneven a2a on this rig;
the byte proofs never hit a trailing short block.

Second: the QSA rows kernel hands its partials over in fp32, so the bf16
wire option inherited fp32 from the input and did nothing. The wire dtype
is now torch.bfloat16 by name on both merge paths.

Tests: test_bar1_a2a_round_offsets.py (2; the skewed 60/20/16 case with
slot 8 checks every round's offset stays inside its block and every byte
moves once), test_dcp_lse_merge.py with fp32 partials (3); suite
tests/moe_offload 537 passed / 43 failed (baseline 43); bar1 unit tests
267 passed / 2 failed (the two abort_431 deadline tests fail identically
at 68c3108c0fce).

+++ tests/moe_offload/test_dcp_lse_merge.py
@@ -65,7 +65,9 @@ def _run(monkeypatch, mode, dtype):
     monkeypatch.setattr(comm, "cp_local_head_bounds", lambda g, c: bounds[g.rank_in_group])
     T, H, D = 7, 6, 4
     w = _World()
-    outs = [torch.randn(T, H, D, dtype=torch.bfloat16) for _ in range(3)]
+    # fp32 partials, as the QSA rows kernel hands them over (fn7j): the bf16
+    # wire must be chosen by the option, not inherited from the input dtype
+    outs = [torch.randn(T, H, D, dtype=torch.float32) for _ in range(3)]
     lses = [torch.randn(T, H) for _ in range(3)]
     for r in range(3):
         w.lses[r] = lses[r]
